ml/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
fastf1==3.4.0
pandas==2.2.3
python-dotenv==1.0.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
pydantic==2.10.5
httpx==0.28.1
